+++ bin/src/calc.dart
@@ -4,6 +4,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:dart_coveralls/dart_coveralls.dart';
+import 'package:path/path.dart' as p;
 
 import 'command_line.dart';
 
@@ -19,30 +20,38 @@ class CalcPart extends CommandLinePart {
           help: "Number of workers for parsing", defaultsTo: "1")
       ..addOption("output", help: "Output file path")
       ..addOption("package-root",
-          help: "Root of the analyzed package", defaultsTo: ".");
+          help: 'Where to find packages, that is, "package:..." imports.',
+          defaultsTo: "packages");
   }
 
   Future execute(ArgResults res) async {
+    String packageRoot = res["package-root"];
+    if (p.isRelative(packageRoot)) {
+      packageRoot = p.absolute(packageRoot);
+    }
+
+    if (!FileSystemEntity.isDirectorySync(packageRoot)) {
+      print("Package root directory does not exist");
+      return;
+    }
+
     if (res.rest.length != 1) {
       print("Please specify a test file to run");
       return;
     }
 
-    var pRoot = new Directory(res["package-root"]);
-    var file = new File(res.rest.single);
-    var workers = int.parse(res["workers"]);
-
-    if (!pRoot.existsSync()) {
-      print("Root directory does not exist");
-      return;
+    var file = res.rest.single;
+    if (p.isRelative(file)) {
+      file = p.absolute(file);
     }
 
-    if (!file.existsSync()) {
+    if (!FileSystemEntity.isFileSync(file)) {
       print("Dart file does not exist");
       return;
     }
 
-    var collector = new LcovCollector(pRoot, file);
+    var workers = int.parse(res["workers"]);
+    var collector = new LcovCollector(packageRoot, file);
 
     var r = await collector.getLcovInformation(workers: workers);